+++ src/screens/MediaScreen.tsx
@@ -12,6 +12,7 @@ import { Image } from "expo-image";
 import { useState, useEffect } from "react";
 import { useLocalSearchParams } from "expo-router";
 import { SafeAreaProvider, SafeAreaView } from "react-native-safe-area-context";
+import styles from "../styles/MediaScreenStyle";
 
 interface Media {
 	type: string;
@@ -263,48 +264,3 @@ export default function MediaScreen() {
 		</SafeAreaProvider>
 	);
 }
-
-const styles = StyleSheet.create({
-	centeredView: {
-		flex: 1,
-		justifyContent: "center",
-		alignItems: "center"
-	},
-	modalView: {
-		margin: 20,
-		backgroundColor: "white",
-		borderRadius: 20,
-		padding: 35,
-
-		shadowColor: "#000",
-		shadowOffset: {
-			width: 0,
-			height: 2
-		},
-		shadowOpacity: 0.25,
-		shadowRadius: 4,
-		elevation: 5
-	},
-	button: {
-		borderRadius: 20,
-		padding: 15,
-		marginTop: 10,
-		elevation: 2
-	},
-	buttonOpen: {
-		backgroundColor: "#F194FF"
-	},
-	buttonClose: {
-		backgroundColor: "#6e6e6e"
-	},
-	textStyle: {
-		color: "white",
-		fontWeight: "bold",
-		textAlign: "center"
-	},
-	modalText: {
-		marginBottom: 15,
-		textAlign: "center"
-	},
-	url: { color: "blue" }
-});
